separate code for python bindings into separate file
